# app/service/user/account/api/grpc/v1/account.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "app/service/user/account/api/grpc/v1/account.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Account"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/account/register": {
      "post": {
        "operationId": "Account_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RegisterResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RegisterReq"
            }
          }
        ],
        "tags": [
          "Account"
        ]
      }
    },
    "/user/account/uid/{steamId}": {
      "get": {
        "operationId": "Account_UID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UIDResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "steamId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Account"
        ]
      }
    },
    "/user/account/{uid}": {
      "get": {
        "operationId": "Account_Info",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1InfoResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Account"
        ]
      },
      "put": {
        "operationId": "Account_ChangeName",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ChangeNameResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "Account"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1ChangeNameResp": {
      "type": "object"
    },
    "v1GetAllUIDResp": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        }
      }
    },
    "v1Info": {
      "type": "object",
      "properties": {
        "steamId": {
          "type": "string",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "firstJoin": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1InfoResp": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/v1Info"
        }
      }
    },
    "v1RegisterReq": {
      "type": "object",
      "properties": {
        "steamId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1RegisterResp": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1UIDResp": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "string",
          "format": "int64"
        }
      }
    }
  }
}